abstracted out the conversation title functions into their own class s… (#167)
* abstracted out the conversatio title functions into their own class so that we can subclass it and easily override those methods as needed.

* updated addapter to use new formatter.

* recycler creats a conversationformatter now.

* removing copyright text.
