+++ layer-atlas/src/main/java/com/layer/atlas/AtlasConversationsRecyclerView.java
@@ -28,6 +28,7 @@ import android.view.View;
 import com.layer.atlas.adapters.AtlasConversationsAdapter;
 import com.layer.atlas.messagetypes.AtlasCellFactory;
 import com.layer.atlas.util.AvatarStyle;
+import com.layer.atlas.util.ConversationFormatter;
 import com.layer.atlas.util.ConversationStyle;
 import com.layer.atlas.util.itemanimators.NoChangeAnimator;
 import com.layer.atlas.util.views.SwipeableItem;
@@ -63,7 +64,7 @@ public class AtlasConversationsRecyclerView extends RecyclerView {
         // Don't flash items when changing content
         setItemAnimator(new NoChangeAnimator());
 
-        mAdapter = new AtlasConversationsAdapter(getContext(), layerClient, picasso);
+        mAdapter = new AtlasConversationsAdapter(getContext(), layerClient, picasso, new ConversationFormatter());
         mAdapter.setStyle(conversationStyle);
         super.setAdapter(mAdapter);
         refresh();
