+++ layer-atlas/src/main/java/com/layer/atlas/util/Util.java
@@ -38,13 +38,11 @@ import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.Locale;
-import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 public class Util {
-    private static final String METADATA_KEY_CONVERSATION_TITLE = "conversationName";
     private static final int TIME_HOURS_24 = 24 * 60 * 60 * 1000;
     private static final SimpleDateFormat DAY_OF_WEEK = new SimpleDateFormat("EEE, LLL dd,", Locale.getDefault());
 
@@ -67,38 +65,6 @@ public class Util {
         manager.setPrimaryClip(clipData);
     }
 
-    public static String getConversationTitle(LayerClient client, Conversation conversation) {
-        return getConversationTitle(client, conversation, conversation.getParticipants());
-    }
-
-    public static String getConversationTitle(LayerClient client, Conversation conversation, Set<Identity> participants) {
-        String metadataTitle = getConversationMetadataTitle(conversation);
-        if (metadataTitle != null) return metadataTitle.trim();
-
-        StringBuilder sb = new StringBuilder();
-        Identity authenticatedUser = client.getAuthenticatedUser();
-        for (Identity participant : participants) {
-            if (participant.equals(authenticatedUser)) continue;
-            String initials = participants.size() > 2 ? getInitials(participant) : Util.getDisplayName(participant);
-            if (sb.length() > 0) sb.append(", ");
-            sb.append(initials);
-        }
-        return sb.toString().trim();
-    }
-
-    public static String getConversationMetadataTitle(Conversation conversation) {
-        String metadataTitle = (String) conversation.getMetadata().get(METADATA_KEY_CONVERSATION_TITLE);
-        if (metadataTitle != null && !metadataTitle.trim().isEmpty()) return metadataTitle.trim();
-        return null;
-    }
-
-    public static void setConversationMetadataTitle(Conversation conversation, String title) {
-        if (title == null || title.trim().isEmpty()) {
-            conversation.removeMetadataAtKeyPath(METADATA_KEY_CONVERSATION_TITLE);
-        } else {
-            conversation.putMetadataAtKeyPath(METADATA_KEY_CONVERSATION_TITLE, title.trim());
-        }
-    }
 
     public static String getInitials(Identity user) {
         String first = user.getFirstName();
